add guard and move files into spec and lib dirs
