+++ levi/Gemfile
@@ -1,0 +1,9 @@
+# A sample Gemfile
+source "https://rubygems.org"
+group :development do
+  gem 'guard'
+  gem 'rb-fsevent'
+  gem 'rb-readline'
+  gem 'guard-rspec', require: false
+end
+
